Sort jpg and RAW files into respective directories while copying
It's not yet clear if this is the most efficient way to go. Likely first 
copying and then sorting is a lot more performant.

Also it expects a "JPG" and "RAW" folder right now, although this is 
still configurable through the config file (fix this).

Also this needs a flag to give it a time range. Currently it just dumps 
everything from A to B.